Win9x: Decide on the name for the Win9x targeting macro
`_WIN32_WINDOWS` was the most compelling alternative. That one used to
be the non-NT equivalent of `_WIN32_WINNT`, as explained in the "Using
the Windows Headers" page from the old Platform SDK documentation. (Not
the current one, which removed any mention of this macro.) As such, it
would have matched the once official way of differentiating between
Windows 95 (0x0400)`, 98 `(0x0410)`, and Me `(0x0500)`, whereas
`_WIN32_WINNT` can only specify either NT 4.0 `(0x0400)` or 2000
`(0x0500)`. This would have actually become relevant with the
multi-monitor APIs used by SDL, which were only introduced in Windows
98.

However, the Win9x build should also be able to dynamically work with
Windows 95, which defeats the point of selecting a specific 9x version.

Part of P0310, funded by Root and Arandui.